integrate with user service

+++ user-service/src/main/java/com/tsys/poc/userservice/controller/UserController.java
@@ -1,5 +1,6 @@
 package com.tsys.poc.userservice.controller;
 
+import com.tsys.poc.userservice.config.JWTValidator;
 import com.tsys.poc.userservice.entity.User;
 import com.tsys.poc.userservice.model.Response;
 import com.tsys.poc.userservice.model.UserInfo;
@@ -15,9 +16,9 @@ import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
-
-import java.util.List;
 import java.util.Optional;
+
+import javax.servlet.http.HttpServletRequest;
 
 @RestController
 @RequestMapping("/userinfo")
@@ -26,21 +27,34 @@ public class UserController {
 	@Autowired
 	private UserService userService;
 
+	@Autowired
+	private JWTValidator jWTValidator;
+
 	@PostMapping("/create")
-	public ResponseEntity<User> create(@RequestBody User user) {
-		userService.add(user);
-		return ResponseEntity.status(HttpStatus.CREATED).body(user);
+	public ResponseEntity<User> create(@RequestBody User user, HttpServletRequest request) {
+		if (!jWTValidator.validateToken(request)) {
+			return new ResponseEntity<User>(HttpStatus.UNAUTHORIZED);
+		}
+		return new ResponseEntity<>(userService.add(user), HttpStatus.OK);
 	}
 
 	@GetMapping("/readAll")
-	public Iterable<User> readAll() {
-		return userService.getAll();
+	public ResponseEntity<Iterable<User>> readAll(HttpServletRequest request) {
+		if (!jWTValidator.validateToken(request)) {
+			return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
+		}
+		return new ResponseEntity<>(userService.getAll(), HttpStatus.OK);
 	}
 
 	@GetMapping("/read/{username}")
-	public ResponseEntity<Response> readByUsername(@PathVariable(name = "username") String username) {
-		Optional<User> user = userService.getById(username);
+	public ResponseEntity<Response> readByUsername(@PathVariable(name = "username") String username,
+			HttpServletRequest request) {
 		Response returnData = new Response();
+		if (!jWTValidator.validateToken(request)) {
+			returnData.setMessage("Not a valid user");
+			return new ResponseEntity<Response>(HttpStatus.UNAUTHORIZED);
+		}
+		Optional<User> user = userService.getById(username);
 		if (user.isPresent()) {
 			returnData.setReturnData(user.get());
 			return ResponseEntity.status(HttpStatus.OK).body(returnData);
@@ -52,8 +66,12 @@ public class UserController {
 	}
 
 	@PutMapping("/update")
-	public ResponseEntity<Response> update(@RequestBody UserInfo userInfo) {
+	public ResponseEntity<Response> update(@RequestBody UserInfo userInfo, HttpServletRequest request) {
 		Response returnData = new Response();
+		if (!jWTValidator.validateToken(request)) {
+			returnData.setMessage("Not a valid user");
+			return new ResponseEntity<Response>(HttpStatus.UNAUTHORIZED);
+		}
 		try {
 			userService.updateUser(userInfo.getUser(), userInfo.getUpdatedUser());
 			returnData.setMessage("Updated Successfully");
@@ -67,8 +85,13 @@ public class UserController {
 	}
 
 	@DeleteMapping("/delete/{username}")
-	public ResponseEntity<Response> deleteByUsername(@PathVariable(name = "username") String username) {
+	public ResponseEntity<Response> deleteByUsername(@PathVariable(name = "username") String username,
+			HttpServletRequest request) {
 		Response returnData = new Response();
+		if (!jWTValidator.validateToken(request)) {
+			returnData.setMessage("Not a valid user");
+			return new ResponseEntity<Response>(HttpStatus.UNAUTHORIZED);
+		}
 		try {
 			userService.deleteById(username);
 			return ResponseEntity.status(HttpStatus.NO_CONTENT).body(returnData);
